+++ latexml_oxide/bin/cortex_worker.rs
@@ -417,6 +417,7 @@ impl LatexmlWorker {
       plane1:                    true,
       hackplane1:                false,
       mathtex:                   self.profile.mathtex,
+      url_style:                 latexml_post::crossref::UrlStyle::File,
       navigationtoc:             None,
       split:                     false,
       split_xpath:               None,